backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
flask-bcrypt
psycopg2-binary
openpyxl
python-dotenv
requests